refactor / clean up removed and combined code
