--- backend/model/ttt.py
@@ -1,10 +1,0 @@
-import passwordClassifer
-def createTokens(f):
-    tokens = []
-    for i in f:
-        tokens.append(i)
-    return tokens
-
-def inputPassword(password):
-    return passwordClassifer.classifyPassword(password)
-
